Move helpers to link_navigation module
